services/agents/app/requirements.txt
# Agent Service Dependencies
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-google-genai
langchain-huggingface
crewai
pinecone-client
sentence-transformers
crawl4ai
requests
